feat(victoria): V247 Phase 2 — wf_obs.py observability instruments (dual-tail report + open-key re-entry coupling counter), validated on V246 ledgers: 320/343 open-key matched, 23 deleted / 7 created, 7/7 creations are re-entries after mechanism-created exits

+++ scripts/v247_ruler.py
@@ -56,7 +56,7 @@ def pctl(sorted_xs: list[float], q: float) -> float:
     if n == 1:
         return sorted_xs[0]
     pos = q / 100.0 * (n - 1)
-    lo = int(math.floor(pos))
+    lo = math.floor(pos)
     hi = min(lo + 1, n - 1)
     frac = pos - lo
     return sorted_xs[lo] * (1 - frac) + sorted_xs[hi] * frac
